+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.4] - 2020-05-16
+
+### Fixed
+
+- Fixed incorrect file and link in progress.log
+- Fixes for legacy windows: Bar, Panel, and Rule now use ASCII characters
+- show_cursor is now a no-op on legacy windows
+
+### Added
+
+- Added Console.input
+
+### Changed
+
+- Disable progress bars when not writing to a terminal
+
 ## [1.1.3] - 2020-05-15
 
 ### Fixed